- Update profile.html to include a button to logout - Update routes.py to include a POST /logout - Remove dead routes and templates - Break router into root and auth files, updating login + register html - Modularize dependencies
