--- app/templates/banking.html
@@ -1,19 +1,0 @@
-{% extends "/shared/_informational.html" %}
-{% block title %}
-    Net Worth
-{% endblock title %}
-{% block topic %}
-    Banking
-{% endblock topic %}
-{% block summary %}
-    One of the first steps to financial independence is maintaining cash reserves that you 
-    can access during an emergency. Generally, it's recommended to have three types of cash
-    accounts:
-{% endblock summary %}
-<h1>ayo</h1>
-{% block goto_link %}
-    /banking
-{% endblock goto_link %}
-{% block goto_title %}
-    Let's Go!
-{% endblock goto_title %}
