+++ app/templates/fragments/invalid-popup.html
@@ -1,0 +1,5 @@
+<div
+    class="flex flex-row justify-between w-1/2 h-16 p-2 m-2 align-middle border rounded-md md:w-1/3 bg-slate-200 border-slate-400">
+    <p class="self-center w-1/2 text-4xl text-center text-yellow-500 md:w-1/3">&#9888</p>
+    <p class="self-center w-full text-sm italic text-left text-black">{{popup_text}}</p>
+</div>
