--- app/templates/income-survey.html
@@ -1,59 +1,0 @@
-{% extends "/shared/_base.html" %} {% block content %}
-<h1 class="h-12 mt-5 text-3xl font-bold text-center">Income</h1>
-<form id="income-form" action="">
-  <!-- gross annual -->
-  <!-- <div class="flex flex-col" id="gross-annual-group">
-    <p
-      class="self-center w-10/12 p-3 overflow-auto text-left rounded-md text-l bg-slate-50"
-    >
-      Gross Annual Income is the money you make in a year before any taxes
-    </p>
-    <div
-      class="self-center w-10/12 p-3 mt-0.5 mb-5 overflow-auto rounded-md bg-slate-50 min-h-1/3"
-    >
-      <label
-        for="gross-annual-income"
-        class="self-center w-screen h-12 text-xl text-black"
-        >Gross Annual Income</label
-      >
-      <input
-        id="gross-annual-income"
-        type="text"
-        class="w-full h-12 p-2 border border-lime-900 focus:outline-lime-500"
-        validationType="currency"
-      />
-    </div>
-  </div> -->
-  <!-- net monthly -->
-  <div class="flex flex-col" id="net-monthly-group">
-    <p
-      class="self-center w-10/12 p-3 overflow-auto text-left rounded-md text-l bg-slate-50"
-    >
-      Net Monthly Income is the money you make in a month after taxes. Often
-      this will be two paychecks!
-    </p>
-    <div
-      class="self-center w-10/12 p-3 mt-0.5 mb-5 overflow-auto rounded-md bg-slate-50 min-h-1/3"
-    >
-      <label
-        for="net-monthly-income"
-        class="self-center w-screen h-12 text-xl text-black"
-        >Net Monthly Income</label
-      >
-      <input
-        id="net-monthly-income"
-        type="text"
-        class="w-full h-12 p-2 border border-lime-900 focus:outline-lime-500"
-        validationType="currency"
-      />
-    </div>
-  </div>
-</form>
-<!-- ensure that user can only supply numbers and decimal point. format as USD after exiting focus of element -->
-<script
-  id="currency-validation"
-  src="{{ url_for('static', path='js/currencyValidations.js') }}"
-></script>
-<!-- <script src="{{ url_for('static', path='js/htmx.min.js') }}"></script> -->
-
-{% endblock content %}
