--- app/templates/income.html
@@ -1,20 +1,0 @@
-{% extends "/shared/_informational.html" %}
-{% block title %}
-    Income
-{% endblock title %}
-{% block topic %}
-    Income
-{% endblock topic %}
-{% block summary %}
-    Your income is the engine of your financial growth! 
-    Using the money you bring in on an annual and monthly basis, we can plan to 
-    spend, save, and invest to meet your goals.
-    <br/><br/>
-    For our next step, let's collect your income information!
-{% endblock summary %}
-{% block goto_link %}
-    /income/survey
-{% endblock goto_link %}
-{% block goto_title %}
-    Let's Go!
-{% endblock goto_title %}
