--- app/templates/net-worth.html
@@ -1,17 +1,0 @@
-{% extends "/shared/_informational.html" %}
-{% block title %}
-    Net Worth
-{% endblock title %}
-{% block topic %}
-    Net Worth
-{% endblock topic %}
-{% block summary %}
-    Your net worth represents the difference between what you OWN and what you OWE. To
-    learn where your net worth stands today, let's get a picture of your account balances.
-{% endblock summary %}
-{% block goto_link %}
-    /banking
-{% endblock goto_link %}
-{% block goto_title %}
-    Let's Go!
-{% endblock goto_title %}
