+++ app/templates/register.html
@@ -4,14 +4,13 @@
 <h1 class="h-12 mt-8 text-3xl font-bold text-center text-green-700">
   Welcome to Financial Forecast
 </h1>
-<form action="/register" method="post" class="flex flex-col items-center justify-center w-screen h-full my-8">
+<!-- TODO: instead of reloading whole page on bad registration, reload just the form html using htmx -->
+<form action="/auth/register" method="post" class="flex flex-col items-center justify-center w-screen h-full my-8">
   {% if popup_text and username %}
-  <div
-    class="flex flex-row justify-between w-1/2 h-16 p-2 m-2 align-middle border rounded-md md:w-1/3 bg-slate-200 border-slate-400">
-    <p class="self-center w-1/2 text-4xl text-center text-yellow-500 md:w-1/3">&#9888</p>
-    <p class="self-center w-full text-sm italic text-left text-black">{{popup_text}}</p>
-  </div>
+  {% include 'fragments/invalid-popup.html' %}
   {% endif %}
+
+  <!-- red outline for invalid username -->
   {% if status == "username_exists" %}
   <input required name="username" value="{{username if username else ''}}" type="text" placeholder="your@email.com"
     autocomplete="username" class="w-1/2 h-12 p-2 border md:w-1/3 border-rose-600 focus:outline-rose-400" />
@@ -19,15 +18,23 @@
   <input required name="username" value="{{username if username else ''}}" type="text" placeholder="your@email.com"
     autocomplete="username" class="w-1/2 h-12 p-2 border md:w-1/3 border-lime-900 focus:outline-lime-500" />
   {% endif %}
+
+  <!-- red outline for invalid password -->
+  {% if status == "bad_password" %}
+  <input required name="password" type="password" placeholder="password" autocomplete="current-password"
+    class="w-1/2 h-12 p-2 mb-6 border md:w-1/3 border-rose-600 focus:outline-rose-400" />
+  {% else %}
   <input required name="password" type="password" placeholder="password" autocomplete="current-password"
     class="w-1/2 h-12 p-2 mb-6 border md:w-1/3 border-lime-900 focus:outline-lime-500" />
+  {% endif %}
   <button type="submit"
     class="self-center w-1/2 h-12 text-xl text-center text-white bg-green-700 border rounded-lg md:w-1/3 border-lime-900 hover:bg-green-600">
     Register
   </button>
+
   <p class="self-center w-1/2 text-center md:w-1/3 mt-7 text-md">
     Already have an account?
   </p>
-  <a href="/login" class="self-center w-1/2 italic text-center text-green-800 md:w-1/3 text-md">Login instead!</a>
+  <a href="/auth/login" class="self-center w-1/2 italic text-center text-green-800 md:w-1/3 text-md">Login instead!</a>
 </form>
 {% endblock %}
