+++ app/templates/shared/_base.html
@@ -1,21 +1,22 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- these were added to appease linter -->
-    <meta name="description" content="foo" />
-    <meta name="keywords" content="foo, bar" />
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', path='css/main.css') }}"
-      type="text/css"
-    />
-    <title>{% block title %}{% endblock %}</title>
-  </head>
-  <body>
-    {% block content %}
-    <h1>Hello World!</h1>
-    {% endblock %}
-  </body>
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <!-- these were added to appease linter -->
+  <meta name="description" content="foo" />
+  <meta name="keywords" content="foo, bar" />
+  <link rel="stylesheet" href="{{ url_for('static', path='css/main.css') }}" type="text/css" />
+  <title>{% block title %}{% endblock %}</title>
+</head>
+
+<body>
+  {% block content %}
+  <h1>Hello World!</h1>
+  {% endblock %}
+  <script src="{{ url_for('static', path='js/htmx.min.js') }}"></script>
+
+</body>
+
 </html>
